Correct SGR code to terminate blinking range
The test-script is incorrectly using SGR 26 to cancel SGR 6, but SGR 5
and SGR 6 are _both_ cancelled by SGR 25. SGR 26 is unrelated.

This fixes the issue that the 'no rapid blink' text is blinking in
Windows Terminal, which implements all of SGR 5, 6 and 25; although it
does not differentiate SGR 5 and SGR 6's blink speed.

References:
* https://en.wikipedia.org/wiki/ANSI_escape_code#SGR_(Select_Graphic_Rendition)_parameters
* Page 61 of https://www.ecma-international.org/wp-content/uploads/ECMA-48_5th_edition_june_1991.pdf
* https://www.real-world-systems.com/docs/ANSIcode.html

Signed-off-by: Paul "TBBle" Hampson <Paul.Hampson@Pobox.com>